platform/x86: int3472: Rework discrete quirks handling
Rework the discrete quirk handling to use a quirks struct which is
pointed to by a dmi_system_id table, rather then having a dmi_system_id
table per type of quirk.

This is a preparation patch for adding support for more quirks, where in
some cases a single model may need multiple quirks and it is undesirable
to have to add multiple dmi_system_id table entries for the same model.

This resolves the quirks before allocating int3472_discrete_device because
future quirks will cause probe() to exit early with -ENODEV.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>