platform/x86: int3472: Allow mapping reset / powerdown functions to regulators
Some designs like the Terra Pad 1262 v2 tablet share the GPIOs which are
marked by the _DSM with reset and powerdown functions between 2 sensors.

Since sharing a GPIO between drivers is not possible at an option to map
these to regulators (with each GPIO providing a separate supply-name)
as a workaround. Regulators can be shared and a regulator will not enter
its disabled state until the last consumer disables it.

Mapping this to regulators is not entirely accurate / pretty but this
allows using unmodified sensor drivers on designs which share these pins
between sensors. Keeping all the ugliness for handling this local to
the INT3472 driver.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>