platform/x86: int3472: Add post() callback to the quirks struct
Some designs are uniquely broken requiring some manual fixup after adding
the clks / regulators / GPIO mappings.

For example on the Terra Pad 1262 v2 the INT3474:01 sensor has a wrong ACPI
I2C resource, causing the i2c-core automatic i2c-client instantiation to
create a client on the wrong bus and address.

This can be worked around by manually creating the i2c-client before
calling acpi_dev_clear_dependencies() as that triggers the instantiation
and if the ACPI fwnode is marked as enumerated it will be skipped.

Add a post() callback to struct int3472_discrete_quirks which can be set
to a manual fixup function which will be run before calling
acpi_dev_clear_dependencies() for cases like this.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>