hid: hid-lg-g15: Use standard multicolor LED API
Replace the custom "color" sysfs attribute with the standard multicolor
LED API.

This also removes the code for the custom "color" sysfs attribute,
the "color" sysfs attribute was never documented so hopefully it is not
used by anyone.

If we get complaints we can re-add the "color" sysfs attribute as
a compatibility wrapper setting the subleds intensity.

Signed-off-by: Kate Hsuan <hpa@redhat.com>
Link: https://lore.kernel.org/r/20241218085955.555316-1-hpa@redhat.com
[hdegoede: various small cleanups + reword commit-msg]
Signed-off-by: Hans de Goede <hdegoede@redhat.com>