media: mt9m114: Add powerdown GPIO support
The actual mt9m114 sensor chip does not have a powerdown pin, but some
camera connectors have both a reset and a powerdown pin and depending on
firmware tables and/or the sensor-module the sensor's reset pin may be
connected to a powerdown GPIO instead.

Both GPIOs may be available, so to make sure that we properly control
the one actually driving the chip's reset pin request and control both
GPIOs to work around this issue.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>