x86/efi/mixed: Set up 1:1 mapping of lower 4GiB in the stub
In preparation for dropping the dependency on startup_32 entirely in the
next patch, add the code that sets up the 1:1 mapping of the lower 4 GiB
of system RAM to the mixed mode stub.

The reload of CR3 after the long mode switch will be removed in a
subsequent patch, when it is no longer needed.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Link: https://lore.kernel.org/r/20250210174941.3251435-13-ardb+git@google.com
Signed-off-by: Hans de Goede <hdegoede@redhat.com>