x86/efi/mixed: Remove dependency on legacy startup_32 code
The EFI mixed mode startup code calls into startup_32() in the legacy
decompressor, passing a mocked up boot_params struct, only to get it to
set up the 1:1 mapping of the lower 4 GiB of memory and switch to a GDT
that supports 64-bit mode.

In order to be able to reuse the EFI mixed mode startup code in EFI boot
images that do not incorporate the legacy decompressor code, decouple
it, by populating the GDT directly.

Doing so allows constructing a GDT that is compatible with the one the
firmware uses, with one additional entry for a 64-bit mode code segment
appended. This removes the need entirely to switch between GDTs and IDTs
or data segment selector values on every call into the firmware, and all
of this code can be removed.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Link: https://lore.kernel.org/r/20250210174941.3251435-14-ardb+git@google.com
Signed-off-by: Hans de Goede <hdegoede@redhat.com>