media: ov02c10: Fix hts for 2 lane mode
Using half the hts value when using 2 lanes results in hts < width,
which results in reporting a negative hblank value to userspace.

The hts value in the mode struct is only used for reporting the hblank
control to userspace so this change does not result in any different
register settings being written to the sensor.

After this change the register-lists are still writing 1140 to the HTS
register of the sensor in 2 lane mode, which seems to be a too low value
for HTS. But maybe the value of this register is multiplied by 2 internally
by the sensor when 2 lanes are used ?

Signed-off-by: Hans de Goede <hdegoede@redhat.com>