media: ov02c10: Fix vts_min for 2 lane mode
Doubling the default VTS value in 2 lane mode to keep the default fps at 30
fps while doubling the pixelrate make sense. But there is no reason to also
double the minimum VTS value. If userspace wants to use the extra bandwidth
to get more fps (at the expense of the max exposure time) then this should
be allowed.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>