i2c: usbio: Fix usbio_i2c_init()
The USBIO_I2CCMD_INIT command does not return any data, so no return buffer
should be passed to usbio_transfer().

This fixes detecting (reading the id register of) the ov08x40 sensor on
a Lenovo ThinkPad X1 carbon gen 12, but there still is an I2C IO error on
stream on, so this still needs further work.

Signed-off-by: Hans de Goede <hansg@kernel.org>