media: mt9m114: Add and use mt9m114_ifp_get_border() helper function
Normally the IFP removes a 4 pixel border all around its sink format
size for demosaicing. But in RAW10 mode it does not do this.

Add a new mt9m114_ifp_get_border() helper function to get the border size
(4 or 0) and use this where applicable instead of hardcoding a border
of 4 pixels everywhere.

Signed-off-by: Hans de Goede <hansg@kernel.org>
---
Changes in v3:
- New patch in v3 of this patch-set