agp/amd64: Change agp_try_unsupported default to 0
AMD64 boards with AGP support are so old that the agp_amd64_pci_table
should be complete and there is no need to probe unlisted PCI devices
by default.

This also brings the driver inline with the global agp=try_unsupported
(agp_try_unsupported_boot) parameter which also default to 0 and brings
things inline with the Kconfig which says that agp_try_unsupported=1
should be passed to try unsupported bridges.

Suggested-by: Lukas Wunner <lukas@wunner.de>
Signed-off-by: Hans de Goede <hansg@kernel.org>