gpio: usbio: Ignore firmware provided valid pins bitmap
One some devices the firmware provided valid pins bitmap is invalid,
so instead of error-ing out when a requested pin is not in the bitmap,
warn about this once and continue.

An example of this is the Lenovo Thinkpad X1 2-in-1 gen 10, where the
request pin for the CVS chip has an offset of 5, where as the bitmap
for the first bank is 0x1f, triggering the bitmap check.

Manually toggeling the pin shows that it is indeed the request pin,
so the problem is with the bitmap reported by the USBIO firmware.

Signed-off-by: Hans de Goede <hansg@kernel.org>