usb: usbio: Rework disconnect/release handling
Rework disconnect/release handling:

1. Disconnect() calls auxiliary_device_delete() on all registered clients,
detaching all the drivers. So there is no need to set client->bridge to
NULL from release() and check for this in functions the aux-device drivers
may call, as all drivers are detached.

Besides dropping the !client->bridge checks also drop checks for drivers
calling functions with NULL pointers, drivers can be trusted to not do
this.

2. OTOH freeing the memory underpinning the aux-device from disconnect()
is too early, since there might still be a reference to it. To fix this
delay freeing the client struct till usbio_auxdev_release() which gets
called when the ref-count drops to 0.

While at it also simplify usbio_gpio_init() a bit.

Signed-off-by: Hans de Goede <hansg@kernel.org>