platform/x86: x86-android-tablets: Switch Acer A1-840 over to generic_lipo_4v2_battery_node
Remove the custom battery-node for the Acer A1-840 and use the generic
generic_lipo_4v2_battery_node instead.

This adds a ocv-capacity-table which is necessary for the new revision
of the chtdc_ti_battery driver which is being upstreamed.

The only change from the now removed custom node is that the max charging
current in the generic node is 2.0A instead of 2.25A, since the current
in the generic node is lower this is not a problem.

Signed-off-by: Hans de Goede <hansg@kernel.org>