platform/x86: x86-android-tablets: Remove Acer A1-840 fuel-gauge sw-node on exit
The Acer A1-840 fuel-gauge sw-node contains a "monitored-battery" reference
which will become invalid when the x86_android_tablets module gets removed.

Remove the sw-node so that if the driver for the fuel-gauge gets re-probed
and tries to parse the device-properties it will not find the properties
instead of deferencing in invalid pointer.

Note that the power-supply-core holds a reference to the "supplied-from"
property (string).

This was strdup-ed by device_create_managed_software_node() so as long
as the sw-node does not gets free-ed (which frees the string) this is ok,
this means that the sw-node must be leaked to avoid a crash.

Signed-off-by: Hans de Goede <hansg@kernel.org>