platform/x86: x86-android-tablets: Unset Acer A1-840 fuel-gauge sw-node on exit
Leaving the software-node added to the DC TI PMIC fuel-gauge platform-dev
around after "rmmod x86_android_tablets" causes 2 issues:

1. device_create_managed_software_node() creates a copy of the properties,
but the "monitored-battery" reference will still become invalid when
the x86_android_tablets module gets removed.

2. When doing a "modprobe x86_android_tablets" after unloading the module,
things fail with an ugly oops because of the chtdc_ti_battery platform-dev
already containing a software_node symlink from the software-node from
the previous device_create_managed_software_node() call.

Switch to manually managing the software-node's life-time and remove it
from an exit() callback to fix both issues.

Note that the power-supply-core holds a reference to the "supplied-from"
property (string) when the software-node is removed.

This was strdup-ed by fwnode_create_software_node(). As long as
the software-node does not gets free-ed (which frees the string) this
is ok, this means that the node must be leaked to avoid a crash.

Signed-off-by: Hans de Goede <hansg@kernel.org>